+++ app/assets/stylesheets/stylesheets/shared/header.css
@@ -112,3 +112,15 @@
   height: 50px;
   background-color: blanchedalmond;
 }
+
+#search-result{
+  width: 30%;
+}
+
+.child{
+  margin: 5px;
+  border: 1px solid #dedede;
+  background-color: #F4F4F4;
+  padding: 5px;
+  font-size: 12px;
+}
