+++ app/javascript/application.js
@@ -3,5 +3,5 @@ import "@hotwired/turbo-rails"
 import "controllers"
 import "item"
 import "card"
-import "name"
 import "category"
+import "item_name"
