+++ app/javascript/category.js
@@ -1,22 +1,14 @@
-document.addEventListener("turbo:load", () => {
+window.addEventListener('load', function(){
   const Categorys = document.getElementById('categorys')
   const Categoryslists = document.getElementById('lists')
   const Categoryslist = document.querySelector('list')
   const Categorysprice = document.getElementById('price')
+  //Categoryslists.style.display = 'none';
   Categorys.addEventListener('click', function() {
-    if (Categoryslist.getAttribute("style") == "display:block;") {
-      Categoryslist.removeAttribute("style", "display:block;")
-    } else {
-      Categoryslist.setAttribute("style")
-    }
-  })
-  Categoryslist.addEventListener('input', function() {
-    Categoryslist.remove();
-    return null;
-  })
-  Categorysprice.addEventListener('input', function() {
-    Categorysprice.remove();
-    return null;
-  })
+    Categoryslists.getAttribute("style","display:block;")
+
+     // cre
+  });
+ 
   
 });
