--- app/javascript/name.js
@@ -1,32 +1,0 @@
-document.addEventListener("turbo:load", () => {
-  const tagNameInput = document.querySelector("#item_name");
-    if (tagNameInput){
-      const inputElement = document.getElementById("item_name");
-      inputElement.addEventListener("input", () => {
-        const keyword = document.getElementById("item_name").value;
-        const XHR = new XMLHttpRequest();
-        XHR.open("GET", `/items/search/?keyword=${keyword}`, true);
-        XHR.responseType = "json";
-        XHR.send();
-        XHR.onload = () => {
-          const searchResult = document.getElementById("search-result");
-          searchResult.innerHTML = "";
-          if (XHR.response) {
-            const tagName = XHR.response.keyword;
-            tagName.forEach((tag) => {
-              const childElement = document.createElement("div");
-              childElement.setAttribute("class", "child");
-              childElement.setAttribute("id", tag.id);
-              childElement.innerHTML = item.name;
-              searchResult.appendChild(childElement);
-              const clickElement = document.getElementById(tag.id);
-              clickElement.addEventListener("click", () => {
-                document.getElementById("item_name").value = clickElement.textContent;
-                clickElement.remove();
-              });
-            });
-          };
-        };
-      });
-    };
-  });
