+++ config/importmap.rb
@@ -7,5 +7,5 @@ pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 pin "item", to: "item.js"
 pin "card", to: "card.js"
-pin "name", to: "name.js"
 pin "category", to: "category.js"
+pin "item_name", to: "item_name.js"
